pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-uv-demo"
version = "0.1.0"
description = "Minimal Python app to test SCA lockfile support (uv, etc)."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Jamie White" }
]
license = {text = "MIT"}
dependencies = [
  "fastapi==0.114.0",
  "uvicorn[standard]==0.30.6",
  "httpx==0.27.0",
  "requests==2.32.3",
  "pydantic==2.8.2",
  "pytest==8.3.2",
  "aiosqlite==0.20.0",
  "SQLAlchemy==2.0.34",
  "redis==5.0.8",
  "cryptography==43.0.1",
  "Jinja2==3.1.4",
  "python-dotenv==1.0.1",
  "loguru==0.7.2",
  "click==8.1.7",
  "typer==0.12.5",
  "boto3==1.35.15"
]

[project.optional-dependencies]
dev = ["pytest==8.3.2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
